% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{barley}
\alias{barley}
\title{Data from a paper by Arild Vold on response of barley to nitrogen fertilizer}
\format{A data frame with 76 rows and 3 columns
\describe{
  \item{year}{Year when the trial was conducted (1970-1988).}
  \item{NF}{Nitrogen fertilizer (g/m^2).}
  \item{yield}{Grain yield of barley (g/m^2).}
}}
\source{
Aril Vold (1998). A generalization of ordinary yield response functions. Ecological Applications. 108:227-236.
}
\usage{
barley
}
\description{
Data from a paper by Arild Vold on response of barley to nitrogen fertilizer
}
\keyword{datasets}
