% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lm.R
\name{boot_lm}
\alias{boot_lm}
\title{Bootstrapping for linear models}
\usage{
boot_lm(object, f = NULL, R = 999, psim = 2,
  resid.type = c("resample", "normal", "wild"), verbose = FALSE, ...)
}
\arguments{
\item{object}{object of class \code{\link[stats]{lm}}}

\item{f}{function to be applied (and bootstrapped), default coef}

\item{R}{number of bootstrap samples, default 999}

\item{psim}{simulation level for \code{\link{simulate_lm}}}

\item{...}{additional arguments to be passed to function \code{\link[boot]{boot}}}
}
\description{
Bootstraping for linear models
}
\details{
This function is inspired by \code{\link[car]{Boot}}, but has more flexibility
in terms of how data are simulated. This function makes multiple copies 
of the original data, so it can be very hungry in terms of memory use, but
I do not believe this to be a big problem given the models we typically fit.
}
\examples{
\donttest{
require(car)
data(barley, package = "nlraa")
## Fit a linear model (quadratic)
fit.lm <- lm(yield ~ NF + I(NF^2), data = barley)

## Bootstrap coefficients by default
fit.lm.bt <- boot_lm(fit.lm)
## Compute confidence intervals
confint(fit.lm.bt, type = "perc")
## Visualize
hist(fit.lm.bt, 1, ci = "perc", main = "Intercept")
hist(fit.lm.bt, 2, ci = "perc", main = "NF term")
hist(fit.lm.bt, 3, ci = "perc", main = "I(NF^2) term")
}

}
