% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{fm0.Theoph.bt}
\alias{fm0.Theoph.bt}
\title{object for confidence bands vignette fm0.Theoph.bt}
\format{
An object of class \sQuote{boot}
\describe{
  \item{fm0.Theoph.bt}{object created in the vignette in chunk \sQuote{Theoph-PB-between-subjects-2}}
}
}
\source{
this package vignette
}
\usage{
fm0.Theoph.bt
}
\description{
object for confidence bands vignette fm0.Theoph.bt
}
\keyword{datasets}
