% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{fm1.P.bt}
\alias{fm1.P.bt}
\title{object for confidence bands vignette fm1.P.bt}
\format{
An object of class \sQuote{boot}
\describe{
  \item{fm1.P.bt}{object created in the vignette in chunk \sQuote{Puromycin-2}}
}
}
\source{
this package vignette
}
\usage{
fm1.P.bt
}
\description{
object for confidence bands vignette fm1.P.bt
}
\keyword{datasets}
