% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{fm2.Lob.bt}
\alias{fm2.Lob.bt}
\title{object for confidence bands vignette fm2.Lob.bt}
\format{
An object of class \sQuote{boot}
\describe{
  \item{fm2.Lob.bt}{object created in the vignette in chunk \sQuote{Loblolly-methods-2}}
}
}
\source{
this package vignette
}
\usage{
fm2.Lob.bt
}
\description{
object for confidence bands vignette fm2.Lob.bt
}
\keyword{datasets}
