% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{fmm1.bt}
\alias{fmm1.bt}
\title{object for confidence bands vignette fmm1.bt}
\format{
An object of class \sQuote{boot}
\describe{
  \item{fmm1.bt}{object created in the vignette in chunk \sQuote{maizeleafext-2}}
}
}
\source{
this package vignette
}
\usage{
fmm1.bt
}
\description{
object for confidence bands vignette fmm1.bt
}
\keyword{datasets}
