% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_nls.R
\name{IC_tab}
\alias{IC_tab}
\title{Information Criteria Table}
\usage{
IC_tab(..., criteria = c("AIC", "AICc", "BIC"), sort = TRUE)
}
\arguments{
\item{...}{model fit objects fitted to the same data}

\item{criteria}{either \sQuote{AIC}, \sQuote{AICc} or \sQuote{BIC}.}

\item{sort}{whether to sort by weights (default to TRUE)}
}
\description{
Information criteria table with weights
}
\note{
The delta and weights are calculated based on the \sQuote{criteria}
}
\seealso{
\code{\link[bbmle]{ICtab}}
}
