% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{lfmc}
\alias{lfmc}
\title{Live fuel moisture content}
\format{
A data frame with 247 rows and 5 variables:
\describe{
  \item{leaf.type}{ -factor- Species for which data was recorded ("Grass E", "Grass W", "M. spinosum", "S. bracteolactus")}
  \item{time}{ -integer- time in days 1-80}
  \item{plot}{ -factor- plot with levels 1-6 (discrete)}
  \item{site}{ -factor- either P ("East") or SR ("West")}
  \item{lfmc}{ -numeric- Live fuel moisture content (percent)}
  \item{group}{grouping for regression}
}
}
\source{
\doi{10.1002/ece3.5543}
}
\usage{
lfmc
}
\description{
Live fuel moisture content
}
\details{
A dataset containing the leaf.type, time, plot, site and lfmc (live fuel mass concentration)
}
\keyword{datasets}
