% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{Lob.bt.pe}
\alias{Lob.bt.pe}
\title{object for confidence bands vignette Lob.bt.pe}
\format{
An object of class \sQuote{boot}
\describe{
  \item{Lob.bt.pe}{object created in the vignette in chunk \sQuote{Loblolly-bootstrap-estimates-1}}
}
}
\source{
this package vignette
}
\usage{
Lob.bt.pe
}
\description{
object for confidence bands vignette Lob.bt.pe
}
\keyword{datasets}
