% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{maizeleafext}
\alias{maizeleafext}
\title{Maize leaf extension rate as a response to temperature}
\format{
A data frame with 10 rows and 2 columns
\describe{
  \item{temp}{Meristem temperature (in Celsius).}
  \item{rate}{Leaf extension rate (relative to 25 degrees).}
}
}
\source{
Walls, W.R., 1971. Role of temperature in the regulation of leaf extension in Zea mays. Nature, 229: 46-47.
}
\usage{
maizeleafext
}
\description{
Data on leaf extension rate as a response to meristem temperature
in maize. The data are re-created liberally from 
Walls, W.R., 1971. Role of temperature in the regulation of leaf extension in Zea mays. Nature, 229: 46-47. 
The data points are not the same as in the original paper. Some additional
points were inserted to fill in the blanks and allow for reasonable
parameter estimates
}
\keyword{datasets}
