% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_nlme.R
\docType{data}
\name{nlraa.env}
\alias{nlraa.env}
\title{Environment to store options and data for nlraa}
\format{
An object of class \code{environment} of length 2.
}
\usage{
nlraa.env
}
\description{
Environment which stores indecies and data for bootstraping mostly
}
\details{
Create an nlraa environment for bootstrapping
}
\keyword{datasets}
