% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.boot.R
\name{print.boot}
\alias{print.boot}
\title{Print an object of class \code{\link[boot]{boot}}}
\usage{
\method{print}{boot}(x, digits = getOption("digits"), index = 1L:ncol(boot.out$t), ...)
}
\arguments{
\item{x}{A bootstrap output object of class \code{\link[boot]{boot}} generated by one of the bootstrap functions.}

\item{digits}{The number of digits to be printed in the summary statistics.}

\item{index}{Indices indicating for which elements of the bootstrap output summary statistics are required.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
This is a copy of boot::print.boot
}
\author{
Brian Ripley with a bug fix by John Nash
}
