% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gnls.R
\name{simulate_gnls}
\alias{simulate_gnls}
\title{Simulate fitted values from an object of class \sQuote{gnls}}
\usage{
simulate_gnls(object, psim = 1, na.action = na.fail,
  naPattern = NULL, ...)
}
\arguments{
\item{object}{object of class \sQuote{gnls}}

\item{psim}{parameter simulation level, 0: for fitted values, 1: for simulation from 
fixed parameters (assuming a fixed vcov matrix), 2: for simulation considering the 
uncertainty in vcov (not implemented yet)}

\item{na.action}{default \sQuote{na.fail}}

\item{naPattern}{missing value pattern}
}
\description{
Simulate gnls based on \sQuote{nlme::predict.gnls} function
}
\details{
It uses function \sQuote{MASS::mvrnorm} to generate new values for the coefficients
of the model using the Variance-Covariance matrix \sQuote{vcov}
}
