% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nlme.R
\name{simulate_nlme}
\alias{simulate_nlme}
\title{Simulate samples from a nonlinear mixed model from fixed effects}
\usage{
simulate_nlme(object, f = NULL, R = 1, psim = 1, ...)
}
\arguments{
\item{object}{object of class \sQuote{gnls} or \sQuote{nlme}}

\item{f}{function to be used for the simulation}

\item{...}{additional arguments (none used at the moment)}

\item{nsim}{number of samples, default 1}
}
\value{
It returns a matrix with simulated values from the original object
with number of rows equal to the number of rows of \sQuote{fitted} and number
of columns equal to the number of simulated samples (\sQuote{nsim}).
}
\description{
Simulate samples from a nonlinear model
}
\details{
details to come
}
\examples{
\donttest{
require(car)
require(nlme)
data(barley)
barley2 <- subset(barley, year < 1974)
fit.lp.gnls2 <- gnls(yield ~ SSlinp(NF, a, b, xs), data = barley2)
barley2$year.f <- as.factor(barley2$year)
cfs <- coef(fit.lp.gnls2)
fit.lp.gnls3 <- update(fit.lp.gnls2, 
                      params = list(a + b + xs ~ year.f),
                      start = c(cfs[1], 0, 0, 0, 
                                cfs[2], 0, 0, 0,
                                cfs[3], 0, 0, 0))
## This will take a few seconds                               
fit.lp.gnls.Bt3 <- boot_nlme(fit.lp.gnls3) 
confint(fit.lp.gnls.Bt3, type = "perc")
}

}
