% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_nlme_one.R
\name{simulate_nlme_one}
\alias{simulate_nlme_one}
\title{Simulate fitted values from an object of class \sQuote{nlme}}
\usage{
simulate_nlme_one(object, psim = 1, level = Q, asList = FALSE,
  na.action = na.fail, naPattern = NULL, ...)
}
\arguments{
\item{object}{object of class \sQuote{gnls}}

\item{psim}{parameter simulation level, 0: for fitted values, 1: for simulation from 
fixed parameters (assuming a fixed vcov matrix), 2: for simulation considering the 
uncertainty in vcov (not implemented yet)}

\item{na.action}{default \sQuote{na.fail}}

\item{naPattern}{missing value pattern}
}
\description{
Simulate nlme based on \sQuote{nlme::predict.nlme} function
}
\details{
It uses function \sQuote{MASS::mvrnorm} to generate new values for the coefficients
of the model using the Variance-Covariance matrix \sQuote{vcov}
}
