% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sm}
\alias{sm}
\title{Sorghum and Maize growth in Greece}
\format{
A data frame with 235 rows and 5 columns
\describe{
  \item{DOY}{ -integer- Day of the year 141-303}
  \item{Block}{ -integer- Block in the experimental design 1-4}
  \item{Input}{ -integer- Input level 1 (Low) or 2 (High)}
  \item{Crop}{ -factor- either F (Fiber Sorghum), M (Maize), S (Sweet Sorghum)}
  \item{Yield}{ -numeric- Biomass yield in Mg/ha}
}
}
\source{
See above reference. (Currently available on ResearchGate).
}
\usage{
sm
}
\description{
Sorghum and Maize growth in Greece
}
\details{
A dataset containing growth data for sorghum and maize in Greece.

Danalatos, N.G.,  S.V. Archontoulis, and K. Tsiboukas. 2009. 
Comparative analysis of sorghum versus corn growing under optimum and under 
water/nitrogen limited conditions in central Greece. In: From research to 
industry and markets: Proceedings of the 17th European Biomass Conference, 
Hamburg, Germany. 29 June–3 July 2009. ETA–Renewable Energies, Florence, 
Italy. p. 538–544.
}
\keyword{datasets}
