% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSbgf4.R
\name{SSbgf4}
\alias{SSbgf4}
\alias{bgf4}
\title{self start for Beta Growth Function with four parameters}
\usage{
bgf4(time, w.max, t.e, t.m, t.b)

SSbgf4(time, w.max, t.e, t.m, t.b)
}
\arguments{
\item{time}{input vector (x) which is normally 'time'.}

\item{w.max}{value of weight or mass at its peak}

\item{t.e}{time at which the weight or mass reaches its peak.}

\item{t.m}{time at which half of the maximum weight or mass has bean reached.}

\item{t.b}{time at which biomass growth starts}
}
\value{
a numeric vector of the same length as x (time) containing parameter estimates for equation specified

vector of the same length as x (time) using the beta growth function with four parameters

a numeric vector of the same length as x (time) containing parameter estimates for equation specified
}
\description{
Self starter for Beta Growth function with parameters w.max, t.m, t.e and t.b
}
\details{
Beta Growth Function

For details see the publication by Yin et al. (2003) "A Flexible Sigmoid Function of Determinate Growth"

Given this function weight is expected to decay and reach zero again at 2*t.e - t.m
}
\examples{
\dontrun{
data(sm)
Examples from old vignette
}
x <- seq(0, 17, by = 0.25)
y <- bgf4(x, 5, 10, 3, 1)
plot(x, y)
y <- bgf4(1:10, 5, 3, 10, 1)
}
