% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSblin.R
\name{SSblin}
\alias{SSblin}
\alias{blin}
\title{self start for Bilinear Function}
\usage{
blin(x, a, b, xs)

SSblin(x, a, b, xs)
}
\arguments{
\item{x}{input vector}

\item{a}{the intercept}

\item{b}{the slope}

\item{xs}{break point of transition between linear and plateau}
}
\value{
a numeric vector of the same length as x containing parameter estimates for equation specified

vector of the same length as x using the bilinear function

a numeric vector of the same length as x containing parameter estimates for equation specified
}
\description{
Self starter for Bilinear function with parameters a (intercept), b (slope), xs (break-point)
}
\details{
This function is described in Archontoulis and Miguez (2015) - (doi:10.2134/agronj2012.0506)
}
\examples{
\dontrun{
require(ggplot2)
set.seed(123)
x <- 1:30
y <- blin(x, 0, 1, 20) + rnorm(30, 0, 0.5)
dat <- data.frame(x = x, y = y)
fit <- nls(y ~ SSblin(x, a, b, xs), data = dat)
## plot
ggplot(data = dat, aes(x = x, y = y)) + 
  geom_point() + 
  geom_line(aes(y = fitted(fit)))
}
}
