% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdlogis.R
\name{SSdlogis}
\alias{SSdlogis}
\alias{dlogis7}
\alias{SSdlogis7}
\alias{dlogis4}
\alias{SSdlogis4}
\title{self start for a double logistic function}
\usage{
dlogis7(x, y.min, y.mid, y.max, xmid1, scal1, xmid2, scal2)

SSdlogis7(x, y.min, y.mid, y.max, xmid1, scal1, xmid2, scal2)

dlogis4(x, yt, yd, xmid, scal)

SSdlogis4(x, yt, yd, xmid, scal)
}
\arguments{
\item{x}{input vector (x). Concentration of substrate in the original Hill model.}

\item{Ka}{parameter representing the concentration at which half of maximum y is attained}

\item{n}{parameter which controls the curvature}

\item{a}{parameter which controls the maximum value of the response (asymptote)}
}
\value{
dlogis7: vector of the same length as x using the quadratic-plateau-quadratic function

dlogis4: vector of the same length as x using the quadratic-plateau-quadratic function
}
\description{
Self starter for Hill function with parameters Ka, n and a
}
\details{
For details see \sQuote{Double logistic curve in regression modeling}

The form of the equations are: \cr
hill1: \deqn{1 / (1 + (Ka/x))}. \cr 
hill2: \deqn{1 / (1 + (Ka/x)^n)}. \cr
hill3: \deqn{a / (1 + (Ka/x)^n)}. \cr
}
\note{
Zero values are not allowed.
}
\examples{
\donttest{
require(ggplot2)
## Example for hill1
set.seed(1234)
x <- 1:20
y <- hill1(x, 10) + rnorm(20, sd = 0.03)
dat1 <- data.frame(x = x, y = y)
fit1 <- nls(y ~ SShill1(x, Ka), data = dat1)

## Example for hill2
y <- hill2(x, 10, 1.5) + rnorm(20, sd = 0.03)
dat2 <- data.frame(x = x, y = y)
fit2 <- nls(y ~ SShill2(x, Ka, n), data = dat2)

## Example for hill3
y <- hill3(x, 10, 1.5, 5) + rnorm(20, sd = 0.03)
dat3 <- data.frame(x = x, y = y)
fit3 <- nls(y ~ SShill3(x, Ka, n, a), data = dat3)

ggplot(data = dat3, aes(x, y)) + 
  geom_point() + 
  geom_line(aes(y = fitted(fit3)))
}
}
