% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSexpf.R
\name{SSexpf}
\alias{SSexpf}
\alias{expf}
\title{self start for exponential unction}
\usage{
expf(x, a, c)

SSexpf(x, a, c)
}
\arguments{
\item{x}{input vector (x)}

\item{a}{represents the value at x = 0}

\item{c}{represents the exponential rate}
}
\value{
a numeric vector of the same length as x containing parameter estimates for equation specified

vector of the same length as x using the profd function

a numeric vector of the same length as x containing parameter estimates for equation specified
}
\description{
Self starter for a simple exponential function
}
\details{
This function is described in Archontoulis and Miguez (2015) - (doi:10.2134/agronj2012.0506) and originally Johnson et al. (2010) Annals of Botany 106: 735–749, 2010. doi:10.1093/aob/mcq183
}
\examples{
\dontrun{
require(ggplot2)
set.seed(1234)
x <- 1:10
y <- expf(x, 10, -0.1) + rnorm(10, 0, 0.2)
dat <- data.frame(x = x, y = y)
fit <- nls(y ~ SSexpf(x, a, c), data = dat)
## plot
ggplot(data = dat, aes(x = x, y = y)) + 
  geom_point() + 
  geom_line(aes(y = fitted(fit)))
}
}
