% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsLMList.R
\name{nlsLMList.formula}
\alias{nlsLMList.formula}
\title{Formula method for nls 'LM' list method}
\usage{
\method{nlsLMList}{formula}(
  model,
  data,
  start = NULL,
  control,
  level,
  subset,
  na.action = na.fail,
  algorithm = c("LM", "default", "port", "plinear"),
  pool = TRUE,
  warn.nls = NA
)
}
\arguments{
\item{model}{see \code{\link{nlsList}}}

\item{data}{see \code{\link{nlsList}}}

\item{start}{see \code{\link{nlsList}}}

\item{control}{see \code{\link{nls}}}

\item{level}{see \code{\link{nlsList}}}

\item{subset}{see \code{\link{nlsList}}}

\item{na.action}{see \code{\link{nlsList}}}

\item{algorithm}{choice of algorithm default is 'LM' which uses 'nlsLM' from the minpack.lm package.}

\item{pool}{see \code{\link{nlsList}}}

\item{warn.nls}{see \code{\link{nlsList}}}
}
\description{
formula method for nlsLMList
}
