% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSdlf.R
\name{SSdlf}
\alias{SSdlf}
\alias{dlf}
\title{self start for Declining Logistic Function}
\usage{
dlf(time, asym, a2, xmid, scal)

SSdlf(time, asym, a2, xmid, scal)
}
\arguments{
\item{time}{input vector (x) which is normally 'time', the smalles value should be close to zero.}

\item{w.max}{value of weight or mass at its peak}

\item{t.m}{time at which half of the maximum weight or mass has bean reached.}

\item{t.e}{time at which the weight or mass reaches its peak.}
}
\value{
a numeric vector of the same length as x (time) containing parameter estimates for equation specified

vector of the same length as x (time) using the declining logistic function

vector of the same length as x (time) using the declining logistic function
}
\description{
Self starter for Beta Growth function with parameters w.max, t.m and t.e
}
\details{
Declining Logistic

Response function:  y = (asym - a2) / (1 + exp((xmid - time)/scal))) + a2 
 .- asym: upper asymptote
 .- xmid: time when y is midway between w and a 
 .- scal: controls the slope
 .- a2: lower asymptote 
For details see the publication by Yin et al. (2003) "A Flexible Sigmoid Function of Determinate Growth"

Given this function weight is expected to decay and reach zero again at 2*t.e - t.m
}
\examples{
\dontrun{
data(sm)
Examples from old vignette
}
x <- seq(0, 17, by = 0.25)
y <- dlf(x, 2, 10, 8, 1)
plot(x, y)
x <- seq(0, 17, by = 0.25)
y <- dlf(x, 2, 10, 8, 1)
plot(x, y)
}
